WIP #52; created separate module for websocket client code
