+++ requirements.txt
@@ -1,0 +1,20 @@
+atomicwrites==1.1.5
+attrs==18.1.0
+coverage==4.5.1
+gevent==1.3.4
+gevent-websocket==0.10.1
+greenlet==0.4.13
+more-itertools==4.2.0
+packaging==17.1
+pluggy==0.6.0
+py==1.5.4
+pyparsing==2.2.0
+pytest==3.6.3
+pytest-cov==2.5.1
+pyzmq==17.1.0
+requests==2.4.3
+six==1.11.0
+tox==3.1.2
+virtualenv==16.0.0
+Werkzeug==0.14.1
+zmq==0.0.0
