+++ tox.ini
@@ -1,17 +1,19 @@
 [tox]
-envlist = py27,py34
+envlist = py34
 
 [testenv]
-deps=
-    py27: mock>=1.3.0
-    six
-    pytest
-    pytest-pudb
-    werkzeug
-;    greenlet==0.4.12
-    gevent-websocket
-    gevent
-    requests==2.4.3
-    zmq
-commands=py.test -rs
-
+;usedevelop = true
+;deps=
+;    six
+;    pytest
+;    pytest-cov
+;    werkzeug
+;    greenlet
+;    gevent-websocket
+;    gevent
+;    requests==2.4.3
+;    zmq
+deps = -rrequirements.txt
+commands=
+    pytest -rs
+;    pytest --cov=tinyrpc/ --cov-report=term --cov-report=html
